+++ src/lib/extension-client.mjs
@@ -903,7 +903,9 @@ export class HybridExtensionClient {
             microflow: options.microflow ?? options.item,
             module: options.module,
             listVariable: options.listVariable ?? options.list ?? options.sourceList,
-            outputVariableName: options.outputVariableName ?? options.outputVariable ?? options.output
+            outputVariableName: options.outputVariableName ?? options.outputVariable ?? options.output,
+            insertBeforeActivity: options.insertBeforeActivity ?? options.insertBefore ?? options.beforeActivity ?? options.beforeCaption,
+            insertBeforeIndex: options.insertBeforeIndex ?? options.beforeIndex
         }), options.timeoutMs);
 
         return {
@@ -932,7 +934,9 @@ export class HybridExtensionClient {
             microflow: options.microflow ?? options.item,
             module: options.module,
             listVariable: options.listVariable ?? options.list ?? options.sourceList,
-            outputVariableName: options.outputVariableName ?? options.outputVariable ?? options.output
+            outputVariableName: options.outputVariableName ?? options.outputVariable ?? options.output,
+            insertBeforeActivity: options.insertBeforeActivity ?? options.insertBefore ?? options.beforeActivity ?? options.beforeCaption,
+            insertBeforeIndex: options.insertBeforeIndex ?? options.beforeIndex
         }), options.timeoutMs);
 
         return {
@@ -962,7 +966,9 @@ export class HybridExtensionClient {
             module: options.module,
             listVariable: options.listVariable ?? options.list ?? options.sourceList,
             objectVariable: options.objectVariable ?? options.value ?? options.itemVariable ?? options.variable,
-            outputVariableName: options.outputVariableName ?? options.outputVariable ?? options.output
+            outputVariableName: options.outputVariableName ?? options.outputVariable ?? options.output,
+            insertBeforeActivity: options.insertBeforeActivity ?? options.insertBefore ?? options.beforeActivity ?? options.beforeCaption,
+            insertBeforeIndex: options.insertBeforeIndex ?? options.beforeIndex
         }), options.timeoutMs);
 
         return {